+++ source/blender/compositor/operations/distort/COM_ScaleOperation.cpp
@@ -48,7 +48,7 @@ ccl_kernel scaleVariableOp(CCL_WRITE(dst),
   READ_IMG(x_input, dst_coords, x_input_pix);
   READ_IMG(y_input, dst_coords, y_input_pix);
 
-  int2 color_coords;
+  float2 color_coords;
   if (relative) {
     color_coords.x = center_x + (dst_coords.x - center_x) / x_input_pix.x;
     color_coords.y = center_y + (dst_coords.y - center_y) / y_input_pix.x;
@@ -79,7 +79,7 @@ ccl_kernel scaleFixedOp(CCL_WRITE(dst),
 
   COORDS_TO_OFFSET(dst_coords);
 
-  int2 input_coords;
+  float2 input_coords;
   if (has_scale_offset) {
     input_coords.x = ((dst_coords.x - scale_offset_x) * scale_rel_x);
     input_coords.y = ((dst_coords.y - scale_offset_y) * scale_rel_y);
